pyproject.toml
[project]
name = "ameath"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pystray>=0.19.0",
]
